#fix #jsx Update log component to dynamically render data and columns using AccountActivitiesModule.
