+++ frontend/src/pages/student/log.jsx
@@ -1,6 +1,8 @@
 import ButtonComponent from "@/components/button/index.jsx";
+import TableComponent from "@/components/table/index.jsx";
 import Title, { Paragraph } from "@/components/title/index.jsx";
 import ColorsContext from "@/context/colors.jsx";
+import { AccountActivitiesModule } from "@/modules/student-account.jsx";
 import { Pagination, Table } from "antd";
 import { useContext, useState } from "react";
 
@@ -11,62 +13,9 @@ export const StudentLog = () => {
   const handleChangePagination = (page) => {
     setCurrentPagination(page);
   };
-  const columns = [
-    {
-      title: "Date/Time",
-      dataIndex: "date",
-      key: "date",
-      render: (date) => <Paragraph fontSize={"text-base"}>{date}</Paragraph>,
-    },
-    {
-      title: "Activity",
-      dataIndex: "activity",
-      key: "activity",
-      align: "center",
-      render: (activity) => (
-        <Paragraph className={"text-center"} fontSize={"text-base"}>
-          {activity}
-        </Paragraph>
-      ),
-    },
-    {
-      title: "Browser",
-      dataIndex: "browser",
-      key: "browser",
-      align: "center",
-      render: (browser) => (
-        <Paragraph className={"text-center"} fontSize={"text-base"}>
-          {browser}
-        </Paragraph>
-      ),
-    },
-    {
-      title: "From IP",
-      dataIndex: "IP",
-      key: "IP",
-      align: "center",
-      render: (IP) => (
-        <Paragraph className={"text-center"} fontSize={"text-base"}>
-          {IP}
-        </Paragraph>
-      ),
-    },
-  ];
 
-  const data = [
-    {
-      date: "Mar 14, 2024 @ 1:29 pm EST",
-      activity: "Student LogOut",
-      browser: "Chrome",
-      IP: "109.207.249.165:38314",
-    },
-    {
-      date: "Mar 14, 2024 @ 1:29 pm EST",
-      activity: "Student LogOut",
-      browser: "Chrome",
-      IP: "109.207.249.165:38314",
-    },
-  ];
+  const { data, columns } = AccountActivitiesModule();
+
   return (
     <div>
       <Title
@@ -95,7 +44,7 @@ export const StudentLog = () => {
         />
       </div>
       <div className={"-mx-5 pt-5"}>
-        <Table columns={columns} dataSource={data} pagination={false} />
+        <TableComponent columns={columns} data={data} />
       </div>
     </div>
   );
